Migrated the eight issues
