+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue2929.cs
@@ -7,8 +7,6 @@ namespace Microsoft.Maui.TestCases.Tests.Issues
 	public class Issue2929 : _IssuesUITest
 	{
 		const string Success = "Success";
-		const string Go = "Go";
-
 		public Issue2929(TestDevice testDevice) : base(testDevice)
 		{
 		}
@@ -17,25 +15,8 @@ namespace Microsoft.Maui.TestCases.Tests.Issues
 
 		[Test]
 		[Category(UITestCategories.ListView)]
-		[Category(UITestCategories.Compatibility)]
-		[FailsOnAndroidWhenRunningOnXamarinUITest]
-		[FailsOnIOSWhenRunningOnXamarinUITest]
-		[FailsOnMacWhenRunningOnXamarinUITest]
 		public void NullItemSourceDoesNotCrash()
 		{
-			// If we can see the Success label, it means we didn't crash. 
-			App.WaitForElement(Success);
-		}
-
-		[Test]
-		[Category(UITestCategories.ListView)]
-		[Category(UITestCategories.Compatibility)]
-		[FailsOnAllPlatformsWhenRunningOnXamarinUITest]
-		public void SettingItemsSourceToNullDoesNotCrash()
-		{
-			App.WaitForElement(Go);
-			App.Tap(Go);
-
 			// If we can see the Success label, it means we didn't crash. 
 			App.WaitForElement(Success);
 		}
