+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue6260.cs
@@ -17,10 +17,6 @@ namespace Microsoft.Maui.TestCases.Tests.Issues
 
 		[Test]
 		[Category(UITestCategories.Layout)]
-		[Category(UITestCategories.Compatibility)]
-		[FailsOnAndroidWhenRunningOnXamarinUITest]
-		[FailsOnMacWhenRunningOnXamarinUITest]
-		[FailsOnWindowsWhenRunningOnXamarinUITest]
 		public void NonAppCompatBasicSwitchTest()
 		{
 			App.WaitForElement(success);
