+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue7338.cs
@@ -6,8 +6,6 @@ namespace Microsoft.Maui.TestCases.Tests.Issues
 {
 	public class Issue7338 : _IssuesUITest
 	{
-		const string Success = "success";
-
 		public Issue7338(TestDevice testDevice) : base(testDevice)
 		{
 		}
@@ -16,14 +14,10 @@ namespace Microsoft.Maui.TestCases.Tests.Issues
 
 		[Test]
 		[Category(UITestCategories.CollectionView)]
-		[Category(UITestCategories.Compatibility)]
-		[FailsOnAndroidWhenRunningOnXamarinUITest]
-		[FailsOnMacWhenRunningOnXamarinUITest]
-		[FailsOnWindowsWhenRunningOnXamarinUITest]
-		public void Issue3273Test()
+		public void Issue7338Test()
 		{
 			// If the instructions are visible at all, then this has succeeded
-			App.WaitForElement(Success);
+			App.WaitForElement("success");
 		}
 	}
 }
