+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue7371.cs
@@ -14,10 +14,6 @@ namespace Microsoft.Maui.TestCases.Tests.Issues
 
 		[Test]
 		[Category(UITestCategories.RefreshView)]
-		[Category(UITestCategories.Compatibility)]
-		[FailsOnAndroidWhenRunningOnXamarinUITest]
-		[FailsOnMacWhenRunningOnXamarinUITest]
-		[FailsOnWindowsWhenRunningOnXamarinUITest]
 		public async Task RefreshingListViewCrashesWhenDisposedTest()
 		{
 			await Task.Delay(500);
